Big changes... * REPLACE wchar_t uses with uint16_t * REMOVE explicit mention of js_val_base from all constructors * ADD more constructors (e.g., string_view and const char*) * STANDARDIZE on "const&" style in ALL constructors * MAKE "pass_thru" handling of UTF-8 in to_string OPTIONAL * ADD js_val "helpers" (operator<< and operator>>) * testing: js_val mechanics, spaceship 3-way comparison, <<operator and formatting throughout * docs UPDATE primary documentation both in .MD file and in sources
Note that this is a release for actual *conforming* C++20 compilers and libraries... which doesn't seem to be a large list currently, even though C++23 is now "old news", and C++26 is being worked on.  Sigh.

The next commit should include an additional set of work-arounds / hacks, largely to deal with the lack of universal support for a *C++17* feature: std::from_chars actually being implemented for float and double data types.